+++ database/migrations/2020_07_03_103531_create_pessoa_tp_pessoa.php
@@ -15,8 +15,8 @@ class CreatePessoaTpPessoa extends Migration
     {
         Schema::create('pessoa_tp_pessoa', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('pessoa_tp_id');
-            $table->unsignedInteger('pessoa_id');
+            $table->unsignedBigInteger('pessoa_tp_id');
+            $table->unsignedBigInteger('pessoa_id');
             $table->timestamps();
             $table->foreign('pessoa_tp_id')->references('id')->on('pessoa_tp')->onDelete('CASCADE');
             $table->foreign('pessoa_id')->references('id')->on('pessoas')->onDelete('CASCADE');
